+++ src/neuroconv/datainterfaces/ophys/minian/miniandatainterface.py
@@ -1,0 +1,59 @@
+from typing import Optional
+
+from pynwb import NWBFile
+from roiextractors.extraction_tools import PathType
+
+from ..basesegmentationextractorinterface import BaseSegmentationExtractorInterface
+
+
+class MinianSegmentationInterface(BaseSegmentationExtractorInterface):
+    """Data interface for MinianSegmentationExtractor."""
+
+    display_name = "Minian Segmentation"
+    associated_suffixes = (".zarr",)
+    info = "Interface for Minian segmentation data."
+
+    @classmethod
+    def get_source_schema(cls) -> dict:
+        source_metadata = super().get_source_schema()
+        source_metadata["properties"]["folder_path"]["description"] = "Path to .zarr output."
+        return source_metadata
+
+    def __init__(self, folder_path: PathType, verbose: bool = True):
+        """
+
+        Parameters
+        ----------
+        folder_path : str or Path
+            Path to .zarr path.
+        verbose : bool, default True
+            Whether to print progress
+        """
+        super().__init__(folder_path=folder_path)
+        self.verbose = verbose
+
+    def add_to_nwbfile(
+        self,
+        nwbfile: NWBFile,
+        metadata: Optional[dict] = None,
+        stub_test: bool = False,
+        stub_frames: int = 100,
+        include_background_segmentation: bool = True,
+        include_roi_centroids: bool = True,
+        include_roi_acceptance: bool = False,
+        mask_type: Optional[str] = "image",  # Literal["image", "pixel", "voxel"]
+        plane_segmentation_name: Optional[str] = None,
+        iterator_options: Optional[dict] = None,
+    ):
+        super().add_to_nwbfile(
+            nwbfile=nwbfile,
+            metadata=metadata,
+            stub_test=stub_test,
+            stub_frames=stub_frames,
+            include_background_segmentation=include_background_segmentation,
+            include_roi_centroids=include_roi_centroids,
+            include_roi_acceptance=include_roi_acceptance,
+            mask_type=mask_type,
+            plane_segmentation_name=plane_segmentation_name,
+            iterator_options=iterator_options,
+        )